Move logic from map-tables to dashboard
